refactor: 이용약관&개인정보처리방침 모달 컨텐츠 공유를 위한 Content 컴포넌트 분리 (#159)
- 온보딩에 사용되는 이용약관 모달과 개인정보처리방침 모달은 각 컨텐츠를 제외하고 동일한 형태의 모달을 사용합니다.
- 별도의 모달(use-policy-modal, privacy-policy-modal)로 선언하기 보다 모달의 형태는 공유해서 쓸 수 있되, content만 교체해서 쓸 수 있도록 구조를 변경하였습니다.
