+++ src/features/onboarding/ui/policy-modal/policy-modal.css.ts
