+++ src/shared/config/index.ts
@@ -1,3 +1,3 @@
 export * from "./company";
 export * from "./experience";
-export * from "./policy";
+export * from "./policy-use-info";
